+++ apps/store/src/blocks/HeaderBlock/HeaderBlock.tsx
@@ -1,8 +1,8 @@
 import { storyblokEditable } from '@storyblok/react'
 import { useMemo } from 'react'
-import { Header } from '@/components/HeaderNew/Header'
-import { HeaderMenu } from '@/components/HeaderNew/HeaderMenu/HeaderMenu'
-import { ShoppingCartMenuItem } from '@/components/HeaderNew/ShoppingCartMenuItem/ShoppingCartMenuItem'
+import { Header } from '@/components/Header/Header'
+import { HeaderMenu } from '@/components/Header/HeaderMenu/HeaderMenu'
+import { ShoppingCartMenuItem } from '@/components/Header/ShoppingCartMenuItem/ShoppingCartMenuItem'
 import type { ExpectedBlockType, SbBaseBlockProps } from '@/services/storyblok/storyblok'
 import { checkBlockType } from '@/services/storyblok/Storyblok.helpers'
 import { type GeneralMenuBlockProps } from './GeneralMenuBlock'
