+++ apps/store/src/blocks/HeaderBlock/SubMenuBlock.tsx
@@ -2,7 +2,7 @@ import * as NavigationMenuPrimitive from '@radix-ui/react-navigation-menu'
 import { storyblokEditable } from '@storyblok/react'
 import clsx from 'clsx'
 import { MinusIcon, PlusIcon } from 'ui'
-import { MenuItemBlock, type MenuItemBlockProps } from '@/blocks/HeaderBlockNew/MenuItemBlock'
+import { MenuItemBlock, type MenuItemBlockProps } from '@/blocks/HeaderBlock/MenuItemBlock'
 import { ButtonNextLink } from '@/components/ButtonNextLink'
 import {
   navigationItem,
@@ -11,11 +11,11 @@ import {
   navigationMenuWrapper,
   navigationSecondaryItem,
   navigationSecondaryList,
-} from '@/components/HeaderNew/Header.css'
-import { NavigationContent } from '@/components/HeaderNew/NavigationContent'
-import { SecondaryNavigationLink } from '@/components/HeaderNew/NavigationLink/NavigationLink'
-import { NavigationTrigger } from '@/components/HeaderNew/NavigationTrigger/NavigationTrigger'
-import { openIcon, closeIcon } from '@/components/HeaderNew/NavigationTrigger/NavigationTrigger.css'
+} from '@/components/Header/Header.css'
+import { NavigationContent } from '@/components/Header/NavigationContent'
+import { SecondaryNavigationLink } from '@/components/Header/NavigationLink/NavigationLink'
+import { NavigationTrigger } from '@/components/Header/NavigationTrigger/NavigationTrigger'
+import { openIcon, closeIcon } from '@/components/Header/NavigationTrigger/NavigationTrigger.css'
 import type { ExpectedBlockType, SbBaseBlockProps } from '@/services/storyblok/storyblok'
 import { filterByBlockType, getLinkFieldURL } from '@/services/storyblok/Storyblok.helpers'
