+++ apps/store/src/components/ForeverPage/ForeverPage.css.ts
@@ -1,7 +1,7 @@
 import { createVar, style } from '@vanilla-extract/css'
 import { minWidth, tokens } from 'ui'
-import { HEADER_HEIGHT_DESKTOP } from '../HeaderNew/Header.css'
-import { HEADER_HEIGHT_MOBILE } from '../HeaderNew/HeaderMenuMobile/HeaderMenuMobile.css'
+import { HEADER_HEIGHT_DESKTOP } from '../Header/Header.css'
+import { HEADER_HEIGHT_MOBILE } from '../Header/HeaderMenuMobile/HeaderMenuMobile.css'
 
 const headerHeight = createVar()
