+++ apps/store/src/components/Header/HeaderMenu/HeaderMenu.tsx
@@ -5,8 +5,8 @@ import Link from 'next/link'
 import { usePathname } from 'next/navigation'
 import { useState, useEffect, useMemo } from 'react'
 import { CrossIcon, Text, xStack } from 'ui'
-import { type HeaderMenuProps } from '@/blocks/HeaderBlockNew/HeaderBlock'
-import { NestedMenuBlock } from '@/blocks/HeaderBlockNew/NestedMenuBlock'
+import { type HeaderMenuProps } from '@/blocks/HeaderBlock/HeaderBlock'
+import { NestedMenuBlock } from '@/blocks/HeaderBlock/NestedMenuBlock'
 import { useProductMetadata } from '@/components/LayoutWithMenu/productMetadataHooks'
 import { useResponsiveVariant } from '@/utils/useResponsiveVariant'
 import { navigationPrimaryList, HeaderMenuDesktop } from '../Header.css'
