+++ apps/store/src/components/Header/NavigationTrigger/NavigationTrigger.css.ts
