+++ apps/store/src/components/Header/NavigationTrigger/NavigationTrigger.tsx
