+++ apps/store/src/components/Header/NavigationTrigger/NavigationTriggerGeneralMenu.tsx
