+++ apps/store/src/components/Header/ShoppingCartMenuItem/ShoppingBagIcon.tsx
