+++ apps/store/src/components/Header/ShoppingCartMenuItem/ShoppingBagIconSvg.tsx
