+++ apps/store/src/components/Header/ShoppingCartMenuItem/ShoppingCartMenuItem.css.ts
