+++ apps/store/src/components/Header/ShoppingCartMenuItem/ShoppingCartMenuItem.tsx
