+++ apps/store/src/components/PaymentConnectPage/Layout.tsx
@@ -4,8 +4,8 @@ import { type ReactNode } from 'react'
 import { Heading, Space, mq, theme } from 'ui'
 import * as GridLayout from '@/components/GridLayout/GridLayout'
 import { LogoHomeLink } from '@/components/LogoHomeLink'
-import { MENU_BAR_HEIGHT_DESKTOP } from '../HeaderNew/Header.css'
-import { MENU_BAR_HEIGHT_MOBILE } from '../HeaderNew/HeaderMenuMobile/HeaderMenuMobile.css'
+import { MENU_BAR_HEIGHT_DESKTOP } from '../Header/Header.css'
+import { MENU_BAR_HEIGHT_MOBILE } from '../Header/HeaderMenuMobile/HeaderMenuMobile.css'
 
 type Props = {
   title?: string
