+++ apps/store/src/features/carDealership/CarBuyerValidationPage.tsx
@@ -3,8 +3,8 @@ import styled from '@emotion/styled'
 import type { Sprinkles } from 'ui/src/theme/sprinkles.css'
 import { Button, Heading, HedvigLogo, Space, Text, mq, theme } from 'ui'
 import * as GridLayout from '@/components/GridLayout/GridLayout'
-import { HEADER_HEIGHT_DESKTOP } from '@/components/HeaderNew/Header.css'
-import { HEADER_HEIGHT_MOBILE } from '@/components/HeaderNew/HeaderMenuMobile/HeaderMenuMobile.css'
+import { HEADER_HEIGHT_DESKTOP } from '@/components/Header/Header.css'
+import { HEADER_HEIGHT_MOBILE } from '@/components/Header/HeaderMenuMobile/HeaderMenuMobile.css'
 import { AttentionCard, CampaignCard } from '@/components/InfoCard/InfoCard'
 import { InputCarRegistrationNumber } from '@/components/InputCarRegistrationNumber/InputCarRegistrationNumber'
 import { InputSelect } from '@/components/InputSelect/InputSelect'
