+++ apps/store/src/features/memberArea/components/Header.tsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled'
 import { mq, theme } from 'ui'
-import { logoWrapper, MENU_BAR_HEIGHT_DESKTOP } from '@/components/HeaderNew/Header.css'
-import { MENU_BAR_HEIGHT_MOBILE } from '@/components/HeaderNew/HeaderMenuMobile/HeaderMenuMobile.css'
+import { logoWrapper, MENU_BAR_HEIGHT_DESKTOP } from '@/components/Header/Header.css'
+import { MENU_BAR_HEIGHT_MOBILE } from '@/components/Header/HeaderMenuMobile/HeaderMenuMobile.css'
 import { LogoHomeLink } from '@/components/LogoHomeLink'
 
 export const Header = () => (
